+++ extensions/bitwarden/src/components/UnlockForm.tsx
@@ -1,4 +1,16 @@
-import { Action, ActionPanel, Clipboard, Form, getPreferenceValues, Icon, showToast, Toast } from "@raycast/api";
+import {
+  Action,
+  ActionPanel,
+  Clipboard,
+  environment,
+  Form,
+  getPreferenceValues,
+  Icon,
+  List,
+  showToast,
+  Toast,
+} from "@raycast/api";
+import path from "path";
 import { useState } from "react";
 import { DebuggingBugReportingActionSection } from "~/components/actions";
 import { LOCAL_STORAGE_KEY } from "~/constants/general";
@@ -60,8 +72,7 @@ const UnlockForm = ({ pendingAction = Promise.resolve() }: UnlockFormProps) => {
       toast.title = "Unlocking vault...";
       const { error: unlockError } = await bitwarden.unlock(password);
       if (unlockError) {
-        if (unlockError instanceof InvalidSessionTokenError && !args?.retryInvalidSessionToken) {
-          await bitwarden.logout({ reason: "Invalid session token" });
+        if (unlockError instanceof InvalidSessionTokenError) {
           return onSubmit({ retryInvalidSessionToken: true });
         }
         return handleUnlockError(unlockError, {
@@ -111,6 +122,15 @@ const UnlockForm = ({ pendingAction = Promise.resolve() }: UnlockFormProps) => {
   if (showPassword) {
     PasswordField = Form.TextField;
     passwordFieldId = "plainPassword";
+  }
+
+  if (isLoading) {
+    const unlockGif = path.join(environment.assetsPath, "unlock.gif");
+    return (
+      <List>
+        <List.EmptyView icon={unlockGif} title="" />
+      </List>
+    );
   }
 
   return (